Create Draft#1
I have already coded the game, but now I have to 1. Make the Game Move Faster. 2. Make the "Game Piece" not a red square (w/ an image) 3. Make the "obstacles" not a green rectangle (an image) BONUS: make the game speed up when the player reaches a certain score. BONUS2: Make the game use keyboard keys instead of buttons.